common/chat, server: refactor, move all conversion functions to common, add tests (#20690)
* Refactor conversion functions
